requirements-plotting.txt
numpy
pandas
matplotlib
SciencePlots
seaborn
colorcet
cmasher
cmcrameri
pillow
scipy
openpyxl
h5py
